% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_anderson.R
\name{calc_anderson_prcp}
\alias{calc_anderson_prcp}
\title{Calculate Anderson et al. Precipitation}
\usage{
calc_anderson_prcp(daily_avgs, closest_dist, days_before = 1, days_after = 2)
}
\arguments{
\item{daily_avgs}{Output from \code{calc_daily_avgs()}}

\item{closest_dist}{Output from \code{calc_closest_dist()}}

\item{days_before}{Number of days before closest distance (default 1)}

\item{days_after}{Number of days after closest distance (default 2)}
}
\value{
Cumulative precipitation from \code{days_before} to \code{days_after}
the date of the closest distance for each geography
}
\description{
Calculate Anderson et al. Precipitation
}
\examples{
harvey_anderson_prcp <- calc_anderson_prcp(
  harvey_daily_avgs,
  harvey_closest_dist
)
}
