% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_daily_avgs.R
\name{calc_daily_avgs}
\alias{calc_daily_avgs}
\title{Calculate Daily Average for Each Geography}
\usage{
calc_daily_avgs(nClimGrid, geography, geoid = "GEOID")
}
\arguments{
\item{nClimGrid}{Output from \code{get_nClimGrid()}}

\item{geography}{\code{sf} object for the area of interest}

\item{geoid}{Unique identifier for geography}
}
\value{
Daily average for all \code{nClimGrid} measures for each
geography
}
\description{
Calculate Daily Average for Each Geography
}
\examples{
harvey_daily_avgs <- calc_daily_avgs(
  harvey_nClimGrid,
  TXLA_tracts
)
}
