% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_outage.R
\name{calc_outage}
\alias{calc_outage}
\title{Calculate Outage Max Precentage and Average Duration}
\usage{
calc_outage(eaglei, state_fips = NULL)
}
\arguments{
\item{eaglei}{Output from \code{get_outage()}}

\item{state_fips}{state fips code to restrict output to}
}
\value{
Maximum precentage and average duration of outage
by county
}
\description{
Calculate Outage Max Precentage and Average Duration
}
\examples{
harvey_outage <- calc_outage(
  harvey_eaglei,
  TXLA_fips
)
}
