% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposur-package.R
\docType{package}
\name{exposur-package}
\alias{exposur}
\alias{exposur-package}
\title{exposur: Download and Process Hurricane Exposure Data}
\description{
Heavily inspired by the 'hurricaneexposure' package, exposur aims to make the process of downloading and processing hurricane exposure data easier
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ryanzomorrodi.github.io/exposur/}
}

}
\author{
\strong{Maintainer}: Ryan Zomorrodi \email{rzomor2@uic.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
