% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eaglei.R
\name{get_eaglei}
\alias{get_eaglei}
\title{Download and Filter Outage Data}
\usage{
get_eaglei(interval, path = NULL)
}
\arguments{
\item{interval}{Dates to get outage data for}

\item{path}{Path to download files into}
}
\value{
Maximum precentage and average duration of outage
by county
}
\description{
Data is downloaded from publically available data published
by \href{https://www.nature.com/articles/s41597-024-03095-5}{Brelsford et al.}
}
\examples{
harvey_eaglei <- get_eaglei(
  interval = interval(harvey_start, harvey_end + days(5)),
  path = "raw-data/eaglei"
)
}
