% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hurdat2.R
\name{get_hurdat2}
\alias{get_hurdat2}
\title{Download HURDAT2 Best Track}
\usage{
get_hurdat2(storm_id, path = NULL)
}
\arguments{
\item{storm_id}{Storm name in all caps, dash, year (EX: "HARVEY-2017")}

\item{path}{Path to download files into}
}
\value{
HURDAT2 best track for a given storm
}
\description{
Data is downloaded from NOAA \href{https://www.nhc.noaa.gov/data/hurdat/}{HURDAT2 HTTPS Server}
}
\examples{
harvey_hurdat2  <- get_hurdat2(
  storm_id = "HARVEY-2017",
  path = "raw-data/hurdat2"
)
}
