% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_closest_dist.R
\name{calc_closest_dist}
\alias{calc_closest_dist}
\title{Calculate Closest Distance and Time}
\usage{
calc_closest_dist(hurdat2, geography, geoid = "GEOID")
}
\arguments{
\item{hurdat2}{Output from \code{get_hurdat2()}}

\item{geography}{\code{sf} object for the area of interest}

\item{geoid}{Unique identifier for geography}
}
\value{
Closest distance and time to track for each geography
}
\description{
Calculate Closest Distance and Time
}
\examples{
harvey_closest_dist <- calc_closest_dist(
  harvey_hurdat2,
  TXLA_tracts
)
}
