% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposur-package.R
\docType{package}
\name{exposur-package}
\alias{exposur}
\alias{exposur-package}
\title{exposur: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ryan Zomorrodi \email{rzomor2@uic.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
