% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eaglei.R
\name{get_eaglei}
\alias{get_eaglei}
\title{Download and Filter Outage Data}
\usage{
get_eaglei(interval, path = NULL)
}
\arguments{
\item{interval}{Dates to get outage data for}

\item{path}{Path to download files into (if not provided,
files will be download into the \code{tempdir()})}
}
\value{
Maximum precentage and average duration of outage
by county
}
\description{
Downloadeds publically available EAGLE-I data published
by \href{https://www.nature.com/articles/s41597-024-03095-5}{Brelsford et al.}
}
\details{
\subsection{Customer Estimates}{

Customer estimates are generated by adding the number of households
(ACS centered on year of storm) and number of establishments (Census SUBI
for year of storm) per county, getting the percentage of households and
establishments per county by state, and multiplying the percentage by the
total number of customers in a county (EIA average over year of storm)
}
}
\examples{
\dontrun{
harvey_eaglei <- get_eaglei(
  interval = interval(harvey_start, harvey_end + days(5)),
  path = "raw-data/eaglei"
)
}
}
