% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nClimGrid.R
\name{get_nClimGrid}
\alias{get_nClimGrid}
\title{Download and Filter Weather Data}
\usage{
get_nClimGrid(interval, path = NULL)
}
\arguments{
\item{interval}{Dates to get weather data for}

\item{path}{Path to download files into (if not provided,
files will be download into the \code{tempdir()})}
}
\value{
\code{stars} raster of climate data over the interval specified
}
\description{
Downloads NOAA \href{https://www.ncei.noaa.gov/products/land-based-station/nclimgrid-daily}{nClimGrid-Daily}
data for the specified interval.
}
\examples{
\dontrun{
harvey_eaglei <- get_eaglei(
  interval = interval(harvey_start, harvey_end + days(5)),
  path = "raw-data/eaglei"
)
}
}
